% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gameFunctions.R
\name{addPieceToBg}
\alias{addPieceToBg}
\title{Add the current piece to the background (because of a collision)}
\usage{
addPieceToBg(values, lambda, nextPiece)
}
\arguments{
\item{values}{reactive values object with current piece and bg information}

\item{lambda}{current lambda value}

\item{nextPiece}{the reactive value object for the next piece (to become the active piece)}
}
\value{
values reactive values object with current piece and bg information
}
\description{
Add the current piece to the background (because of a collision)
}
\author{
Jason Waddell
}

