% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gameFunctions.R
\name{adjacentLocations}
\alias{adjacentLocations}
\title{Find adjacent locations to a given block}
\usage{
adjacentLocations(pieceRow)
}
\arguments{
\item{pieceRow}{One row of the current piece data.frame}
}
\value{
matrix of adjacent locations
}
\description{
Find adjacent locations to a given block
}
\author{
Jason Waddell
}

