% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gameFunctions.R
\name{clearLines}
\alias{clearLines}
\title{Check if any lines need to be cleared, and update objects accordingly}
\usage{
clearLines(values)
}
\arguments{
\item{values}{reactive values object with current piece and bg information}
}
\value{
(numeric) number of lines cleared
}
\description{
Check if any lines need to be cleared, and update objects accordingly
}
\author{
Jason Waddell
}

