% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gameFunctions.R
\name{detectCollision}
\alias{detectCollision}
\title{Detects whether a proposed change in the active piece will cause a collision in
the background}
\usage{
detectCollision(values, xChange = 0, yChange = 0, rotateClockwise = FALSE,
  rotateCounter = FALSE)
}
\arguments{
\item{values}{reactive values object with current piece and bg information}

\item{xChange}{proposed x location change}

\item{yChange}{proposed y location change}

\item{rotateClockwise}{(logical) whether to rotate the piece clockwise}

\item{rotateCounter}{(logical) whether to rotate the piece counter clockwise}
}
\value{
(logical) whether a collision will occur
}
\description{
Detects whether a proposed change in the active piece will cause a collision in
the background
}
\author{
Jason Waddell
}

