% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gameFunctions.R
\name{findOpenLocations}
\alias{findOpenLocations}
\title{When generating a new piece, find candidate block locations}
\usage{
findOpenLocations(pieceMat)
}
\arguments{
\item{pieceMat}{current piece data.frame}
}
\value{
list of candidate locations
}
\description{
When generating a new piece, find candidate block locations
}
\author{
Jason Waddell
}

