% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gameFunctions.R
\name{makeBg}
\alias{makeBg}
\title{Create a blank background list}
\usage{
makeBg()
}
\value{
the background list
}
\description{
Create a blank background list
}
\author{
Jason Waddell
}

