% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gameFunctions.R
\name{makeShape}
\alias{makeShape}
\title{Make a randomly generated piece, sampled from the current lambda value}
\usage{
makeShape(lambda = 3.5)
}
\arguments{
\item{lambda}{value of lambda (to be handed to rpois)}
}
\value{
a piece list object
}
\description{
Make a randomly generated piece, sampled from the current lambda value
}
\author{
Jason Waddell
}

