% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{plotBg}
\alias{plotBg}
\title{Plot the background list object}
\usage{
plotBg(bg)
}
\arguments{
\item{bg}{list containing x, y and color values}
}
\value{
none (objects plotted to current device)
}
\description{
Plot the background list object
}
\author{
Jason Waddell
}

