% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{plotDotplot}
\alias{plotDotplot}
\title{Make a plot displaying number of blocks in piece versus lambda}
\usage{
plotDotplot(record)
}
\arguments{
\item{record}{the record reactive value object}
}
\value{
none (objects plotted to current device)
}
\description{
Make a plot displaying number of blocks in piece versus lambda
}
\author{
Jason Waddell
}

