% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{plotShape}
\alias{plotShape}
\title{Plot the current shape list object}
\usage{
plotShape(shape, xOffset, yOffset)
}
\arguments{
\item{shape}{shape list object containing x, y and color values}

\item{xOffset}{current x offset modifying shape object}

\item{yOffset}{current y offset modifying shape object}
}
\value{
none (objects plotted to current device)
}
\description{
Plot the current shape list object
}
\author{
Jason Waddell
}

