% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gameFunctions.R
\name{rotateClockwise}
\alias{rotateClockwise}
\title{Rotate a piece clockwise}
\usage{
rotateClockwise(shape)
}
\arguments{
\item{shape}{the shape list object}
}
\value{
updated shape list object
}
\description{
Rotate a piece clockwise
}
\author{
Jason Waddell
}

