% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gameFunctions.R
\name{rotateCounter}
\alias{rotateCounter}
\title{Rotate a piece counterclockwise}
\usage{
rotateCounter(shape)
}
\arguments{
\item{shape}{the shape list object}
}
\value{
updated shape list object
}
\description{
Rotate a piece counterclockwise
}
\author{
Jason Waddell
}

