% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/runPoissontris.R
\name{runPoissontris}
\alias{runPoissontris}
\title{Run the Cardiovascular MFR shiny application}
\usage{
runPoissontris()
}
\value{
no return value
}
\description{
Run the Cardiovascular MFR shiny application
}

